Moved Profile to types package
