+++ types/api.go
@@ -121,3 +121,6 @@ func NewFollowRequest(r io.Reader) (req FollowRequest, err error) {
 	err = json.Unmarshal(body, &req)
 	return
 }
+
+type ProfileResponse struct {
+}
