+++ types/profile.go
@@ -1,4 +1,4 @@
-package internal
+package types
 
 type Profile struct {
 	Type string
